Moving UI creation logic out of EDUS2View.
